Update NavEntryComponentGetterGenerator to use requireRoute (#216)
## Description

Follow up to #212 to update codegen to use renamed method.

In current version of kotlinpoet `Bundle.requireRoute` and `Fragment.requireRoute` clash and cause missing import. To avoid this use import alias for `Bundle` version. This is fixed already in kotlinpoet upstream and could be removed with next release.

Also move `NavEntryComponentGetterGenerator` tests to individual tests together with others to replicate the clash.